pyproject.toml
[project]
name = "financeagents"
version = "0.1.0"
description = "多智能体协作金融问答系统"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.4.8",
    "streamlit>=1.36.0",
    "pydantic>=2.6.0",
    "langchain-openai>=0.3.23",
    "langchain-core>=0.2.0",
    "langchain-community>=0.2.0",
    "pandas>=2.3.0",
    "yfinance>=0.2.63",
    "stockstats>=0.6.5",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.14.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.5",
    "matplotlib>=3.8.3",
    "httpx>=0.27.0",
] 